requirements-dev.txt
requests == 2.28.2
pynacl == 1.5.0
boto3 == 1.26.103
moto[awslambda,dynamodb] == 4.1.6
docker == 6.0.1